// fiverr-clone/src/types/Gigs.ts
export interface Gigs {
  _id: string;
  title: string;
  userId: number;
  cover: string;
  pp: string;
  images: string[];
  desc: string;
  price: number;
  star: number;
  username: string;
  shortDesc: string;
  deliveryTime: number;
  revisionsNumber: number;
  shortTitle: string;
  totalStars: number | 0;
  starNumber: number | 0;
  features: string[];
}